Output files - dtype uint16

Former-commit-id: 6638525b9f24230703b6d445c69a80f28657f242

+++ tomopy/util/dtype.py
@@ -100,7 +100,7 @@ def as_int32(arr):
 
 def as_uint16(arr):
     arr = as_ndarray(arr, np.uint16)
-    return as_dtype(arr, np.int32)
+    return as_dtype(arr, np.uint16)
 
 
 def as_uint8(arr):
